Vulnerability CVE-2022-41946
pgjdbc is an open source postgresql JDBC Driver. In affected versions a
prepared statement using either `PreparedStatement.setText(int,
InputStream)` or `PreparedStatemet.setBytea(int, InputStream)` will
create a temporary file if the InputStream is larger than 2k. This will
create a temporary file which is readable by other users on Unix like
systems, but not MacOS. On Unix like systems, the system's temporary
directory is shared between all users on that system. Because of this,
when files and directories are written into this directory they are, by
default, readable by other users on that same system. This vulnerability
does not allow other users to overwrite the contents of these
directories or files. This is purely an information disclosure
vulnerability. Because certain JDK file system APIs were only added in
JDK 1.7, this this fix is dependent upon the version of the JDK you are
using. Java 1.7 and higher users: this vulnerability is fixed in 4.5.0.
Java 1.6 and lower users: no patch is available. If you are unable to
patch, or are stuck running on Java 1.6, specifying the java.io.tmpdir
system environment variable to a directory that is exclusively owned by
the executing user will mitigate this vulnerability.